pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-screener"
version = "0.1.0"
description = "Distributed stock screening pipeline using Ray"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34",
    "paramiko>=3.4",
    "ray[default]>=2.9",
    "redis>=5.0",
    "marshmallow>=3.21",
    "dacite>=1.8",
    "yfinance>=0.2.36",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pytest-mock>=3.12",
    "moto[ec2,elasticache]>=5.0",
    "fakeredis>=2.21",
]

[project.scripts]
stock-screener = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-n auto --tb=short"
